requirements.txt
torch
Flask
numpy
matplotlib
requests
PyYAML
minio
cvxopt
tqdm
pandas
Pillow
torchvision
python-mnist
tensorboard